+++ src/components/AccordionSection/AccordionSection.test.js
@@ -19,4 +19,32 @@ describe("AccordionSection ", () => {
     );
     expect(wrapper).toMatchSnapshot();
   });
+
+  it("can handle click events on the title", () => {
+    const onTitleClick = jest.fn();
+    let expanded = null;
+    const wrapper = shallow(
+      <AccordionSection
+        content={<span>Test</span>}
+        expanded={expanded}
+        onTitleClick={onTitleClick}
+        setExpanded={id => {
+          expanded = id;
+        }}
+        title="Test section"
+      />
+    );
+    wrapper
+      .find(".p-accordion__tab")
+      .at(0)
+      .simulate("click");
+    expect(onTitleClick.mock.calls[0][0]).toBe(true);
+    wrapper.setProps({ expanded });
+    // Clicking the title again should close the accordion section.
+    wrapper
+      .find(".p-accordion__tab")
+      .at(0)
+      .simulate("click");
+    expect(onTitleClick.mock.calls[1][0]).toBe(false);
+  });
 });